+++ internal/cluster/clusterpb/MetaStateSnapshot.go
@@ -174,8 +174,28 @@ func (rcv *MetaStateSnapshot) IcebergTablesLength() int {
 	return 0
 }
 
+func (rcv *MetaStateSnapshot) ObjectIndex(obj *MetaObjectIndexEntry, j int) bool {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(18))
+	if o != 0 {
+		x := rcv._tab.Vector(o)
+		x += flatbuffers.UOffsetT(j) * 4
+		x = rcv._tab.Indirect(x)
+		obj.Init(rcv._tab.Bytes, x)
+		return true
+	}
+	return false
+}
+
+func (rcv *MetaStateSnapshot) ObjectIndexLength() int {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(18))
+	if o != 0 {
+		return rcv._tab.VectorLen(o)
+	}
+	return 0
+}
+
 func MetaStateSnapshotStart(builder *flatbuffers.Builder) {
-	builder.StartObject(7)
+	builder.StartObject(8)
 }
 func MetaStateSnapshotAddNodes(builder *flatbuffers.Builder, nodes flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(0, flatbuffers.UOffsetT(nodes), 0)
@@ -214,6 +234,12 @@ func MetaStateSnapshotAddIcebergTables(builder *flatbuffers.Builder, icebergTabl
 	builder.PrependUOffsetTSlot(6, flatbuffers.UOffsetT(icebergTables), 0)
 }
 func MetaStateSnapshotStartIcebergTablesVector(builder *flatbuffers.Builder, numElems int) flatbuffers.UOffsetT {
+	return builder.StartVector(4, numElems, 4)
+}
+func MetaStateSnapshotAddObjectIndex(builder *flatbuffers.Builder, objectIndex flatbuffers.UOffsetT) {
+	builder.PrependUOffsetTSlot(7, flatbuffers.UOffsetT(objectIndex), 0)
+}
+func MetaStateSnapshotStartObjectIndexVector(builder *flatbuffers.Builder, numElems int) flatbuffers.UOffsetT {
 	return builder.StartVector(4, numElems, 4)
 }
 func MetaStateSnapshotEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {